feat: hook으로 전환
